Lerna structure + websocket client
