+++ lerna.json
@@ -1,0 +1,7 @@
+{
+  "lerna": "2.9.0",
+  "packages": [
+    "packages/*"
+  ],
+  "version": "1.0.0"
+}
