+++ packages/tcp/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "dat-daemon",
+  "version": "1.0.8",
+  "description": "Dat daemon",
+  "main": "index.js",
+  "directories": {
+    "bin": "bin",
+    "lib": "lib",
+    "test": "test"
+  },
+  "bin": {
+    "datdaemond": "./bin/daemon.js",
+    "datdaemon": "./bin/cli.js"
+  },
+  "scripts": {
+    "test": "standard && tape test/*.js"
+  },
+  "keywords": [
+    "dat"
+  ],
+  "author": "soyuka",
+  "license": "MIT",
+  "dependencies": {
+    "dat-link-resolve": "^2.1.0",
+    "dat-node": "^3.5.8",
+    "level": "^2.1.2",
+    "minimist": "^1.2.0",
+    "mkdirp": "^0.5.1",
+    "prettier-bytes": "^1.0.4",
+    "dat-daemon-protocol": "^1.0.0"
+  },
+  "devDependencies": {
+    "rimraf": "^2.6.2",
+    "standard": "^10.0.3",
+    "tape": "^4.8.0"
+  }
+}
