+++ packages/ws/index.js
@@ -1,0 +1,2 @@
+console.warning('Please require the dat-daemon module instead.')
+module.exports = require('dat-daemon')
