--- ws/cli.js
@@ -1,6 +1,0 @@
-/**
- * @TODO:
- * 1. move original bin/cli.js getMessage to lib/ + export
- * 2. reuse the getMessage with a ws client instead
- * 3. add a bin entry to package.json with this client
- */
