--- ws/index.js
@@ -1,2 +1,0 @@
-console.warning('Please use the dat-daemon module.')
-module.exports = {}
